Update zfarm.ts
PR to update zFarm for u1 to account for the changes in the last few patches. 
Have only updated the typescript file as that seems to be the base file that you use and I'm not sure what system you use to convert it to the contents of the .js file so figured you'd be best doing that.
Added support for the following things:
Devastation, Domination, Frigid, Experience challenges.
Angelic, herbalist, kerfluffle masteries and gamma burst
Fixed Fused C2s not working

One thing that's missing is that I included Randimps in the imp calculations and the input field for that  seems to have a max input of 5 so needs increased to 6.